backend/pyproject.toml
[project]
name = "framefinder-backend"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "channels>=4.3.1",
    "daphne>=4.2.1",
    "django>=5.2.6",
    "django-cors-headers>=4.9.0",
    "djangorestframework>=3.16.1",
    "ffmpeg-python>=0.2.0",
    "langchain-google-genai>=2.1.12",
    "langgraph>=0.6.7",
    "numpy>=2.3.3",
    "openai>=1.109.1",
    "openai-whisper>=20250625",
    "pandas>=2.3.2",
    "python-dotenv>=1.1.1",
    "sentence-transformers>=5.1.1",
]
